Python/requirements.txt
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
PyJWT==2.6.0
pymongo==3.12.3